backend/requirements.txt
cohere==5.16.3
Flask==3.1.1
python-dotenv==1.1.1
requests==2.32.4
spotipy==2.25.1